requirements.txt
absl-py==1.3.0
accelerate==0.20.3
aiohttp==3.8.3
aiosignal==1.3.1
albumentations==1.3.1
altair==5.0.1
aniso8601==9.0.1
antlr4-python3-runtime==4.8
asttokens==2.1.0
async-timeout==4.0.2
attrs==22.1.0
autofaiss==2.15.6
backcall==0.2.0
backports.zoneinfo==0.2.1
blinker==1.5
braceexpand==0.1.7
cachetools==5.2.0
certifi==2023.5.7
charset-normalizer==2.1.1
click==8.1.3
-e git+https://github.com/openai/CLIP.git@d50d76daa670286dd6cacf3bcd80b5e4823fc8e1#egg=clip
clip-anytorch==2.5.0
clip-retrieval==2.35.1
coloredlogs==15.0.1
comm==0.1.3
commonmark==0.9.1
Cython==0.29.32
dataclasses==0.6
debugpy==1.6.7
decorator==5.1.1
diffusers==0.16.1
docker-pycreds==0.4.0
einops==0.3.0
embedding-reader==1.5.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
ExifRead-nocycle==3.0.1
faiss-cpu==1.7.3
filelock==3.8.0
fire==0.4.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
flatbuffers==22.10.26
frozenlist==1.3.3
fsspec==2022.1.0
ftfy==6.1.1
future==0.18.2
gitdb==4.0.9
GitPython==3.1.29
google-auth==2.14.1
google-auth-oauthlib==0.4.6
grpcio==1.50.0
h5py==3.7.0
huggingface-hub==0.15.1
humanfriendly==10.0
idna==3.4
imageio==2.31.1
img2dataset==1.35.0
importlib-metadata==6.7.0
importlib-resources==5.12.0
ipykernel==6.23.2
ipython==8.12.1
ipywidgets==8.0.5
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.0
jupyter_client==8.2.0
jupyter_core==5.3.1
jupyterlab-widgets==3.0.7
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mpmath==1.2.1
multidict==6.0.2
multilingual-clip==1.0.10
nest-asyncio==1.5.6
networkx==2.8.8
nltk==3.7
numpy==1.22.3
oauthlib==3.2.2
omegaconf==2.1.1
onnx==1.12.0
onnxruntime==1.13.1
open-clip-torch==2.7.0
opencv-python==4.6.0.66
opencv-python-headless==4.6.0.66
openpyxl==3.0.10
packaging==21.3
pandas==1.5.1
pandas-stubs==1.5.1.221024
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
platformdirs==2.5.4
prometheus-client==0.15.0
promise==2.3
prompt-toolkit==3.0.32
protobuf==3.20.1
psutil==5.9.4
ptyprocess==0.7.0
pudb==2019.2
pure-eval==0.2.2
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydeck==0.8.0
pydensecrf @ git+https://github.com/lucasb-eyer/pydensecrf.git@0d53acbcf5123d4c88040fe68fbb9805fc5b2fb9
pyDeprecate==0.3.1
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
pytorch-lightning==1.4.2
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
qudida==0.0.4
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rich==12.6.0
rsa==4.9
sacremoses==0.0.53
safetensors==0.3.1
scikit-image==0.19.3
scikit-learn==1.1.3
scipy==1.9.3
semver==2.13.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
sentry-sdk==1.11.1
setproctitle==1.3.2
shortuuid==1.0.11
six==1.16.0
smmap==5.0.0
stack-data==0.6.1
streamlit==1.15.0
sympy==1.11.1
-e git+https://github.com/CompVis/taming-transformers.git@3ba01b241669f5ade541ce990f7650a3b8f65318#egg=taming_transformers
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.1.1
test-tube==0.7.5
threadpoolctl==3.1.0
tifffile==2022.10.10
tokenizers==0.12.1
toml==0.10.2
toolz==0.12.0
torch==1.11.0
torch-fidelity==0.3.0
torchaudio==0.11.0
torchmetrics==0.10.3
torchvision==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
transformers==4.30.2
types-pytz==2022.6.0.1
typing_extensions==4.6.3
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
urwid==2.1.2
validators==0.20.0
wandb==0.12.21
watchdog==2.1.9
wcwidth==0.2.5
webdataset==0.1.103
Werkzeug==2.2.2
widgetsnbextension==4.0.3
yarl==1.8.1
zipp==3.10.0
